requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
black==23.7.0
Brotli==1.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
coverage==7.2.7
cryptography==41.0.3
defusedxml==0.7.1
dj-config-url==0.1.1
dj-database-url==2.0.0
Django==4.2.3
django-allauth==0.54.0
django-cors-headers==4.2.0
django-filter==23.2
django-redis==5.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
ecdsa==0.18.0
exceptiongroup==1.1.2
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
platformdirs==3.6.0
pluggy==1.2.0
psycopg2-binary==2.9.6
pycparser==2.21
Pygments==2.15.1
PyJWT==2.8.0
pytest==7.4.0
pytest-django==4.5.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.23.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
types-PyYAML==6.0.12.10
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
whitenoise==6.5.0